extract HTTP-interface-related code into separate file
