+++ src/canisters/frontend/ic-certified-assets/src/lib.rs
@@ -1,4 +1,5 @@
 //! This module declares canister methods expected by the assets canister client.
+pub mod http;
 pub mod rc_bytes;
 pub mod state_machine;
 pub mod types;
@@ -9,6 +10,7 @@ mod tests;
 
 pub use crate::state_machine::StableState;
 use crate::{
+    http::{HttpRequest, HttpResponse, StreamingCallbackHttpResponse, StreamingCallbackToken},
     rc_bytes::RcBytes,
     state_machine::{AssetDetails, CertifiedTree, EncodedAsset, State},
     types::*,
@@ -233,7 +235,7 @@ fn is_authorized() -> Result<(), String> {
     STATE.with(|s| {
         s.borrow()
             .is_authorized(&caller())
-            .then(|| ())
+            .then_some(())
             .ok_or_else(|| "Caller is not authorized".to_string())
     })
 }
