+++ src/canisters/frontend/ic-certified-assets/src/url_decode.rs
@@ -20,7 +20,7 @@ fn convert_percent(iter: &mut std::slice::Iter<u8>) -> Option<u8> {
     Some(result)
 }
 
-#[derive(Debug, PartialEq)]
+#[derive(Debug, PartialEq, Eq)]
 pub enum UrlDecodeError {
     InvalidPercentEncoding,
 }
